Add style selector and comment checkbox to the text editor pane
Also, make some changes to how styles are stored, so we can track changes to them in order to be able to update the style selector efficiently.

+++ src/nde/mod.rs
@@ -20,7 +20,7 @@ pub struct Event {
     pub start: subtitle::StartTime,
     pub duration: subtitle::Duration,
     pub layer_index: i32,
-    pub style_index: i32,
+    pub style_index: usize,
     pub margins: subtitle::Margins,
 
     /// Tags applying to the entire line.
